feat(live-preview): add AOT mode
Adds an AOT mode to the live-preview.
This no longer allows editing.
