+++ projects/live-preview/components/si-live-preview/si-live-preview.component.html
@@ -240,6 +240,9 @@
       [disabled]="editorCollapsed"
     >
       <div class="buttons" [class.show-copied]="showCopied">
+        @if (!enableJitMode) {
+          <span class="readonly-label">read-only</span>
+        }
         @if (!isMobile) {
           <button
             type="button"
