+++ projects/live-preview/esbuild-component-loader.mjs
@@ -4,6 +4,16 @@ import { globIterate } from 'glob';
 const livePreviewComponentLoader = {
   name: 'live-preview-component-loader',
   setup: build => {
+    // Loader for .html files to return raw string content
+    build.onLoad({ filter: /\.html$/ }, async args => {
+      const fs = await import('fs/promises');
+      const contents = await fs.readFile(args.path, 'utf8');
+      return {
+        contents: `export default ${JSON.stringify(contents)};`,
+        loader: 'js'
+      };
+    });
+
     build.onResolve({ filter: /(@siemens|@simpl)\/live-preview\/component-loader.*/ }, args => {
       const [data] = args.path.split('!');
       const url = new URL('fake:' + data);
@@ -33,18 +43,23 @@ const livePreviewComponentLoader = {
         let code = 'export default { load: function (path) { switch (path) {';
         const files = [];
         const webcomponentsFiles = [];
+
         for await (const file of globIterate(
-          path.posix.join(root, examples.replace(/\.ts$/, '.{ts,tsx,vue,js}')),
+          path.posix.join(root, examples.replace(/\.ts$/, '.{ts,tsx,vue,js,html}')),
           { posix: true } // needed for windows
         )) {
           if (file.endsWith('.ts')) {
             const fileName = file.replace(new RegExp('^' + root + '/'), '').replace(/\.ts$/, '');
             code += `case '${fileName}': return import('${file.replace(/\.ts$/, '')}');`;
             files.push(fileName);
+          } else if (file.endsWith('.html')) {
+            const fileName = file.replace(new RegExp('^' + root + '/'), '');
+            code += `case '${fileName}': return import('${file}');`;
           } else {
             webcomponentsFiles.push(file.replace(root + '/', '').replace(/\.(tsx|vue|js)$/, ''));
           }
         }
+
         files.sort();
         webcomponentsFiles.sort();
