+++ projects/live-preview/public-api.ts
@@ -11,6 +11,7 @@ export * from './components/si-example-viewer/si-example-viewer.component';
 export * from './components/si-live-preview/si-live-preview.component';
 export * from './components/si-live-preview-iframe/si-live-preview-iframe.component';
 export * from './components/si-live-preview-renderer/si-live-preview-renderer.component';
+export * from './components/si-live-preview-renderer/si-live-preview-jit-renderer.component';
 export * from './components/si-live-preview-wrapper/si-live-preview-wrapper.component';
 export * from './components/si-live-preview-qr/si-live-preview-qr.component';
